+++ Projects/todo_list/todo_list_backend/src/main/java/com/todoList/dao/image/ImageDAO.java
@@ -3,7 +3,7 @@ package com.todoList.dao.image;
 import com.todoList.entities.Image;
 
 public interface ImageDAO {
-    Image get(int id);
+    Image get(long id);
     Image save(Image image);
-    void delete(int id);
+    void delete(long id);
 }
