+++ Projects/todo_list/todo_list_backend/src/main/java/com/todoList/dao/todo/TodoDAO.java
@@ -9,9 +9,9 @@ import java.util.List;
 public interface TodoDAO {
     List<Todo> getAllLimited(int from, int to);
     List<Todo> getAll();
-    Todo get(int id) throws NotFoundException;
+    Todo get(long id) throws NotFoundException;
     Todo save(Todo todo);
     Todo update(Todo todo) throws NotFoundException;
-    void updateStatus(int id, Status status);
-    Todo delete(int id) throws NotFoundException;
+    void updateStatus(long id, Status status);
+    Todo delete(long id) throws NotFoundException;
 }
