+++ Projects/todo_list/todo_list_backend/src/main/java/com/todoList/entities/Image.java
@@ -17,8 +17,7 @@ public class Image {
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private int id;
-
+    private long id;
     @NotNull
     private String name;
     @NotNull
