+++ Projects/todo_list/todo_list_backend/src/main/java/com/todoList/entities/Todo.java
@@ -21,8 +21,7 @@ public class Todo
 {
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private int id;
-
+    private long id;
     @NotNull
     @Column(length = 5000)
     private String whatTodo;
