+++ Projects/todo_list/todo_list_backend/src/main/java/com/todoList/services/todo/TodoService.java
@@ -9,9 +9,9 @@ import java.util.List;
 public interface TodoService {
     List<Todo> getAllLimited(int from, int to);
     List<Todo> getAll();
-    Todo get(int id);
+    Todo get(long id);
     Todo save(TodoRequest todo);
     Todo update(Todo todo);
-    void updateStatus(int id, Status status);
-    Todo delete(int id);
+    void updateStatus(long id, Status status);
+    Todo delete(long id);
 }
