+++ Projects/todo_list/todo_list_backend/src/main/java/com/todoList/services/user/UserService.java
@@ -7,7 +7,7 @@ import com.todoList.entities.User;
 public interface UserService {
     User save(User user);
     User getByEmail(String email) throws UnauthorizedNotFoundException;
-    User getById(int id);
+    User getById(long id);
     User update(UserRequest user) throws Exception;
     Boolean checkIfExists(String email);
     void delete() throws Exception;
